+++ app/store/index.ts
@@ -1,3 +1,4 @@
 export * from "./useSettingsStore";
+export * from "./useDebugConsoleStore";
 export * from "./useTaleStore";
 export * from "./useUpdateStore";